requirements.txt
httpx>=0.27.0
PyYAML>=6.0.1,<7.0.0
numpy>=1.24.0,<3.0.0
scipy>=1.10.0,<2.0.0
